+++ src/main/java/edu/harvard/iq/dataverse/UserNotificationServiceBean.java
@@ -6,6 +6,7 @@
 
 package edu.harvard.iq.dataverse;
 
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.logging.Logger;
@@ -67,7 +68,7 @@ public class UserNotificationServiceBean {
             return new Long("0");
         }
         Query query = em.createNativeQuery("select count(id) from usernotification as o where o.user_id = " + userId + " and o.readnotification = 'false';");
-        return (Long) query.getSingleResult();    
+        return (Long) ((BigInteger)query.getSingleResult()).longValue();
     }
     
     public List<UserNotification> findUnemailed() {
